package/genimage: bump version
This commit bumps the genimage version to a not yet released commit,
in order to benefit from the GPT support that was added in upstream
genimage.

Indeed, we are about to add a defconfig for the STM32MP157 Discovery
Kit platform and the STM32MP157 BootROM loads the first stage
bootloader from the GPT partitions named fsbl*, which requires a GPT
partition table.

We will of course be able to switch back to a released version once a
release that includes GPT support will have been done upstream.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>